moved asymmetric key file to separate class
