+++ packages/at_chops/lib/src/key/aes_key.dart
@@ -1,5 +1,6 @@
 import 'package:at_chops/src/key/at_encryption_key.dart';
 
+/// Represents an AES key for symmetric encryption.
 class AESKey extends AtEncryptionKey {
   late String _aesKey;
   AESKey(this._aesKey);
