+++ packages/at_chops/lib/src/key/at_private_key.dart
@@ -1,3 +1,4 @@
+/// Represents a private key from [AtKeyPair]
 abstract class AtPrivateKey {
   AtPrivateKey(String atPrivateKey);
 }
